requirements.txt
pipecat-ai[cartesia,daily,deepgram,elevenlabs,openai,playht,silero]
python-dotenv
aiohttp
fastapi
uvicorn